requirements.txt
openai>=1.55.3,<2
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.0.0
